cleaning up comments, utils, tracing
