+++ src/Sarif.Viewer.VisualStudio.ResultSources.DeveloperCanvas.UnitTests/UtilTest.cs
@@ -38,23 +38,5 @@ namespace Sarif.Viewer.VisualStudio.ResultSources.DeveloperCanvas.UnitTests
             string output = Util.S(text, count);
             output.Should().Be(expected);
         }
-
-        /// <summary>
-        /// Tests that we can properly parse git urls into the server-project-repo format we use.
-        /// </summary>
-        /// <param name="repoUrl">The url of the repo we need to parse.</param>
-        /// <param name="expectedServer">The server we expect to get out.</param>
-        /// <param name="expectedProject">The project we expect to get out.</param>
-        /// <param name="expectedRepo">The repo we expect to get out.</param>
-        [Theory]
-        [InlineData("https://dev.azure.com/serverName/projectName/_git/repoName", "dev.azure.com/serverName", "projectName", "repoName")]
-        [InlineData("https://serverName.visualstudio.com/projectName/_git/repoName", "serverName.visualstudio.com", "projectName", "repoName")]
-        public void ParseGitUrl(string repoUrl, string expectedServer, string expectedProject, string expectedRepo)
-        {
-            Util.ParseGitUrl(repoUrl, out string serverName, out string projectName, out string repoName);
-            serverName.Should().Be(expectedServer);
-            projectName.Should().Be(expectedProject);
-            repoName.Should().Be(expectedRepo);
-        }
     }
 }
